refactor: Move parser functions to runtime_parser.go
